+++ src/main/java/roomescape/web/api/MemberReservationController.java
@@ -11,18 +11,20 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.domain.reservation.Reservation;
+import roomescape.application.CancelService;
 import roomescape.application.ReservationService;
 import roomescape.application.dto.request.member.MemberInfo;
 import roomescape.application.dto.request.reservation.ReservationRequest;
 import roomescape.application.dto.request.reservation.UserReservationRequest;
 import roomescape.application.dto.response.reservation.ReservationResponse;
 import roomescape.application.dto.response.reservation.UserReservationResponse;
+import roomescape.domain.reservation.Reservation;
 
 @RestController
 @RequiredArgsConstructor
 public class MemberReservationController {
     private final ReservationService reservationService;
+    private final CancelService cancelService;
 
     @PostMapping("/reservations")
     public ResponseEntity<ReservationResponse> makeReservation(
@@ -52,7 +54,7 @@ public class MemberReservationController {
             @PathVariable(value = "idWaiting") Long waitingId,
             MemberInfo memberInfo
     ) {
-        reservationService.cancelReservation(waitingId, memberInfo);
+        cancelService.cancelReservation(waitingId, memberInfo);
         return ResponseEntity.noContent().build();
     }
 
@@ -61,7 +63,7 @@ public class MemberReservationController {
             @PathVariable(value = "idReservation") Long reservationId,
             MemberInfo memberInfo
     ) {
-        reservationService.cancelReservation(reservationId, memberInfo);
+        cancelService.cancelReservation(reservationId, memberInfo);
         return ResponseEntity.noContent().build();
     }
 }
